Refactored App.js content
